set cSupplied test
